requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
pandas==2.1.1
numpy==1.25.2 